Updated file structure to clean up code and define locations
